requirements_extended.txt
# Core speech recognition dependencies (already installed)
SpeechRecognition==3.14.4
PyAudio==0.2.14
pyttsx3==2.99

# For Discord bot integration
discord.py==2.3.2

# For Telegram bot integration
pyTelegramBotAPI==4.14.0

# For WhatsApp/Web integration
Flask==3.0.0
Flask-SocketIO==5.3.6
requests==2.31.0

# Additional useful packages
python-dotenv==1.0.0  # For environment variables
websockets==12.0      # For real-time communication
aiohttp==3.9.1        # For async HTTP requests